requirements.txt
tinydb
RPi.GPIO
dynaconf
paho.mqtt